+++ src/components/DoAllocate/DoAllocate.tsx
@@ -4,6 +4,7 @@
 import React, { FC, useEffect, useMemo, useState } from 'react';
 import { gql, useLazyQuery } from '@apollo/client';
 import IPFSImage from '@components/IPFSImage';
+import { useMakeNotification } from '@components/NotificationCentre/useMakeNotification';
 import { NumberInput } from '@components/NumberInput';
 import { NETWORK_NAME, useAccount } from '@containers/Web3';
 import { parseEther } from '@ethersproject/units';
@@ -51,6 +52,7 @@ const DoAllocate: FC<IProps> = ({ projectId, deploymentId, actionBtn, onSuccess,
   const [estimatedRewardsOneEra, setEstimatedRewardsOneEra] = useState(BigNumber(0));
   const [addOrRemove, setAddOrRemove] = useState<'Add' | 'Remove'>(initialStatus || 'Add');
   const waitTransactionHandled = useWaitTransactionhandled();
+  const { refreshAndMakeOverAllocateNotification } = useMakeNotification();
 
   const [fetchTotalDeploymentAllocation, totalDeploymentAllocation] = useLazyQuery(gql`
     query GetDeploymentAllocationSummary($deploymentId: String!) {
@@ -192,6 +194,7 @@ const DoAllocate: FC<IProps> = ({ projectId, deploymentId, actionBtn, onSuccess,
       // runnerAllocation fetch from contract, so if the transaction confirmed, then the data should be updated
       await Promise.all([waitTransactionHandled(receipt?.blockNumber), runnerAllocation.refetch()]);
       await onSuccess?.();
+      refreshAndMakeOverAllocateNotification();
       form.resetFields();
       openNotification({
         type: 'success',
