+++ src/components/NotificationCentre/index.tsx
@@ -1,39 +1,21 @@
 // Copyright 2020-2022 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import React, { FC, useCallback, useEffect, useRef, useState } from 'react';
-import { GoAlertFill, GoBell } from 'react-icons/go';
+import React, { FC, useEffect, useState } from 'react';
+import { GoBell } from 'react-icons/go';
 import { IoIosAlert, IoIosClose } from 'react-icons/io';
 import { useNavigate } from 'react-router';
-import { gql, useLazyQuery } from '@apollo/client';
-import { useSQToken } from '@containers/SQToken';
 import { useAccount } from '@containers/Web3';
 import { useEra } from '@hooks';
-import { useConsumerHostServices } from '@hooks/useConsumerHostServices';
-import { useEthersSigner } from '@hooks/useEthersProvider';
-import { getTotalStake } from '@hooks/useSortedIndexer';
 import { Typography } from '@subql/components';
-import {
-  formatEther,
-  useGetFilteredDelegationsLazyQuery,
-  useGetIndexerLazyQuery,
-  useGetRewardsLazyQuery,
-} from '@subql/react-hooks';
-import { convertBigNumberToNumber, formatSQT } from '@utils';
-import { limitContract, makeCacheKey } from '@utils/limitation';
-import { waitForSomething } from '@utils/waitForSomething';
-import { Badge, Button, Modal, Popover } from 'antd';
-import BigNumberJs from 'bignumber.js';
+import { Badge, Button, Popover } from 'antd';
 import dayjs from 'dayjs';
 
-import { useWeb3Store } from 'src/stores';
-import {
-  type NotificationItem as NotificationItemType,
-  NotificationKey,
-  useNotification,
-} from 'src/stores/notification';
+import { type NotificationItem as NotificationItemType, useNotification } from 'src/stores/notification';
 
 import styles from './index.module.less';
+import { useMakeNotification } from './useMakeNotification';
+import { useToastNotificationModal } from './useToastNotificationModal';
 
 const EmptyNotification = () => {
   return (
@@ -49,449 +31,6 @@ const EmptyNotification = () => {
       </Typography>
     </div>
   );
-};
-
-// TODO: split to other files
-export const useToastNotificationModal = () => {
-  const notificationStore = useNotification();
-  const navigate = useNavigate();
-  const toasting = useRef(false);
-  const toastedKeys = useRef<string[]>([]);
-
-  const toastRestNotifications = useCallback(
-    async (notification?: NotificationItemType) => {
-      if (toasting.current) {
-        await waitForSomething({
-          func: () => toasting.current === false,
-        });
-      }
-      const canBeToastNotification = (notification ? [notification] : notificationStore.notificationList).filter(
-        (item) => {
-          return (
-            (!item.dismissTo || (item.dismissTo && item.dismissTo < Date.now())) &&
-            !toastedKeys.current.includes(item.key)
-          );
-        },
-      );
-
-      let continueShowToastStatus: 'continue' | 'stop' | 'pending' = 'pending';
-      toasting.current = true;
-      for (const noti of canBeToastNotification) {
-        continueShowToastStatus = 'pending';
-        toastedKeys.current.push(noti.key);
-        Modal.confirm({
-          icon: null,
-          width: 376,
-          className: styles.notificationModal,
-          content: (
-            <div className={styles.notificationModalContent}>
-              <GoAlertFill
-                style={{ color: noti.level === 'critical' ? 'var(--sq-error)' : 'var(--sq-info)', fontSize: 40 }}
-              />
-              <Typography weight={600} variant="large">
-                {noti.title}
-              </Typography>
-              {noti.message ? <Typography style={{ textAlign: 'center' }}>{noti.message}</Typography> : ''}
-            </div>
-          ),
-          cancelButtonProps: {
-            style: {
-              display: noti.canBeDismissed ? 'block' : 'none',
-            },
-            shape: 'round',
-            size: 'large',
-          },
-          cancelText: 'Dismiss',
-          okText: noti.buttonProps.label,
-          okButtonProps: {
-            size: 'large',
-            type: 'primary',
-            shape: 'round',
-            danger: noti.level === 'critical' ? true : false,
-          },
-          // eslint-disable-next-line no-loop-func
-          onOk: () => {
-            if (!noti.buttonProps.navigateHref) {
-              continueShowToastStatus = 'continue';
-              return;
-            }
-
-            continueShowToastStatus = 'stop';
-            // window.open may be block by browser
-            if (
-              !noti.buttonProps.navigateHref.includes('https://') &&
-              !noti.buttonProps.navigateHref.includes('http://')
-            ) {
-              navigate(noti.buttonProps.navigateHref);
-            } else {
-              window.open(noti.buttonProps.navigateHref, '_blank');
-            }
-          },
-          // eslint-disable-next-line no-loop-func
-          onCancel: () => {
-            continueShowToastStatus = 'continue';
-            if (noti.dismissTime) {
-              const dismissTo = Date.now() + noti.dismissTime;
-              notificationStore.updateNotification({
-                ...noti,
-                dismissTo,
-              });
-            }
-          },
-        });
-
-        await waitForSomething({
-          // eslint-disable-next-line no-loop-func
-          func: () => {
-            return continueShowToastStatus !== 'pending';
-          },
-          splitTime: 500,
-        });
-
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore
-        if (continueShowToastStatus === 'stop') {
-          notificationStore.removeNotification(noti);
-          return;
-        }
-      }
-      toasting.current = false;
-    },
-    [notificationStore.notificationList],
-  );
-
-  return {
-    toastRestNotifications: toastRestNotifications,
-    toastOneNotification: async (notification: NotificationItemType) => {
-      await toastRestNotifications(notification);
-    },
-  };
-};
-
-const idleTimeout = (func: () => void) => setTimeout(func, 200);
-const idleCallback = window.requestIdleCallback || idleTimeout;
-
-export const useMakeNotification = () => {
-  const notificationStore = useNotification();
-  const { currentEra } = useEra();
-  const { contracts } = useWeb3Store();
-  const { provider } = useEthersSigner();
-  const { account } = useAccount();
-  const [fetchRewardsApi] = useGetRewardsLazyQuery();
-  const [fetchIndexerData] = useGetIndexerLazyQuery();
-  const [fetchIndexerController] = useLazyQuery<{ indexer?: { controller?: string } }>(gql`
-    query Indexer($address: String!) {
-      indexer(id: $address) {
-        controller
-      }
-    }
-  `);
-  const [fetchUnlockWithdrawal] = useLazyQuery<{ withdrawls: { totalCount?: number } }>(gql`
-    query MyQuery($address: String!, $startTime: Datetime!) {
-      withdrawls(
-        filter: {
-          delegator: { equalTo: $address }
-          status: { equalTo: ONGOING }
-          startTime: { lessThanOrEqualTo: $startTime }
-        }
-      ) {
-        totalCount
-      }
-    }
-  `);
-  const { consumerHostBalance } = useSQToken();
-  const { getHostingPlanApi } = useConsumerHostServices({
-    autoLogin: false,
-  });
-
-  // TODO: filter inactive by Graphql
-  const [fetchDelegations] = useGetFilteredDelegationsLazyQuery();
-
-  const makeOverAllocateAndUnStakeAllocationNotification = useCallback(async () => {
-    // over and unused share same api, so must query both at the same time
-    // TODO: Maybe can optimise
-
-    if (
-      notificationStore.notificationList.find((item) => item.key === NotificationKey.OverAllocate) &&
-      notificationStore.notificationList.find((item) => item.key === NotificationKey.UnstakeAllocation) &&
-      notificationStore.notificationList.find((item) => item.key === NotificationKey.OverAllocateNextEra)
-    ) {
-      return;
-    }
-    const res = await contracts?.stakingAllocation.runnerAllocation(account || '');
-    const runnerAllocation = {
-      used: formatSQT(res?.used.toString() || '0'),
-      total: formatSQT(res?.total.toString() || '0'),
-    };
-
-    const isOverAllocated = +runnerAllocation?.used > +runnerAllocation?.total;
-    const isUnused = +runnerAllocation.total - +runnerAllocation.used > 1000;
-    if (
-      isUnused &&
-      !notificationStore.notificationList.find((item) => item.key === NotificationKey.UnstakeAllocation)
-    ) {
-      // add a notification to inform user that they have unused allocation
-      notificationStore.addNotification({
-        key: 'unstakeAllocation',
-        level: 'info',
-        message: `You have ${+runnerAllocation.total - +runnerAllocation.used} SQT unused`,
-        title: 'Unused Allocation',
-        createdAt: Date.now(),
-        canBeDismissed: true,
-        dismissTime: 1000 * 60 * 60 * 24,
-        dismissTo: undefined,
-        type: '',
-        buttonProps: {
-          label: 'Unstake Allocation',
-          navigateHref: '/indexer/my-projects',
-        },
-      });
-    }
-    if (
-      isOverAllocated &&
-      !notificationStore.notificationList.find((item) => item.key === NotificationKey.OverAllocate)
-    ) {
-      notificationStore.addNotification({
-        key: 'overAllocate',
-        level: 'critical',
-        message: `You have used ${runnerAllocation.used} SQT out of ${runnerAllocation.total} SQT allocated`,
-        title: 'Over Allocation',
-        createdAt: Date.now(),
-        canBeDismissed: true,
-        dismissTime: 1000 * 60 * 60 * 24,
-        dismissTo: undefined,
-        type: '',
-        buttonProps: {
-          label: 'Increase Allocation',
-          navigateHref: '/indexer/my-projects',
-        },
-      });
-      notificationStore.sortNotificationList();
-    }
-
-    if (!notificationStore.notificationList.find((item) => item.key === NotificationKey.OverAllocateNextEra)) {
-      const indexerData = await fetchIndexerData({
-        variables: {
-          address: account || '',
-        },
-        fetchPolicy: 'network-only',
-      });
-      if (indexerData.data?.indexer?.id) {
-        const totalStake = getTotalStake(indexerData.data.indexer.totalStake, currentEra.data?.index);
-
-        if (totalStake.after && BigNumberJs(totalStake.after).lt(runnerAllocation.used)) {
-          // add notification to inform user that they may over allocated next era
-          notificationStore.addNotification({
-            key: 'overAllocateNextEra',
-            level: 'info',
-            message: `You have used ${runnerAllocation.used} SQT out of ${totalStake.after} SQT allocated in the next era`,
-            title: 'Over Allocation Next Era',
-            createdAt: Date.now(),
-            canBeDismissed: true,
-            dismissTime: 1000 * 60 * 60 * 24,
-            dismissTo: undefined,
-            type: '',
-            buttonProps: {
-              label: 'Adjust Allocation',
-              navigateHref: '/indexer/my-projects',
-            },
-          });
-        }
-      }
-    }
-  }, [account, contracts, notificationStore.notificationList, currentEra.data?.index]);
-
-  const makeUnClaimedNotification = useCallback(async () => {
-    if (notificationStore.notificationList.find((item) => item.key === NotificationKey.UnclaimedRewards)) {
-      return;
-    }
-    const res = await fetchRewardsApi({
-      variables: {
-        address: account || '',
-      },
-      fetchPolicy: 'network-only',
-    });
-
-    if (res.data?.unclaimedRewards?.totalCount) {
-      notificationStore.addNotification({
-        key: NotificationKey.UnclaimedRewards,
-        level: 'info',
-        message: '',
-        title: 'Unclaimed Rewards',
-        createdAt: Date.now(),
-        canBeDismissed: true,
-        dismissTime: 1000 * 60 * 60 * 24,
-        dismissTo: undefined,
-        type: '',
-        buttonProps: {
-          label: 'Claim Rewards',
-          navigateHref: '/profile/rewards',
-        },
-      });
-    }
-  }, [account, notificationStore.notificationList]);
-
-  const makeLowBillingBalanceNotification = useCallback(async () => {
-    if (notificationStore.notificationList.find((item) => item.key === NotificationKey.LowBillingBalance)) {
-      return;
-    }
-
-    const hostingPlan = await getHostingPlanApi({ account: account || '' });
-    if (!hostingPlan?.data?.length) {
-      return;
-    }
-
-    const res = await consumerHostBalance?.refetch();
-    const [billingBalance] = res?.data || [];
-    if (
-      !BigNumberJs(formatSQT(billingBalance?.toString() || '0')).isZero() &&
-      BigNumberJs(formatSQT(billingBalance?.toString() || '0')).lt(400)
-    ) {
-      notificationStore.addNotification({
-        key: NotificationKey.LowBillingBalance,
-        level: 'critical',
-        message:
-          'Your Billing account balance is running low. Please top up your Billing account promptly to avoid any disruption in usage.',
-        title: 'Low Billing Balance',
-        createdAt: Date.now(),
-        canBeDismissed: true,
-        dismissTime: 1000 * 60 * 60 * 24,
-        dismissTo: undefined,
-        type: '',
-        buttonProps: {
-          label: 'Add Balance',
-          navigateHref: '/consumer/flex-plans/ongoing',
-        },
-      });
-      notificationStore.sortNotificationList();
-    }
-  }, [account, notificationStore.notificationList]);
-
-  const makeInactiveOperatorNotification = useCallback(async () => {
-    if (notificationStore.notificationList.find((item) => item.key === NotificationKey.InactiveOperator)) return;
-    const res = await fetchDelegations({
-      variables: { delegator: account ?? '', filterIndexer: account ?? '', offset: 0 },
-      fetchPolicy: 'network-only',
-    });
-
-    if (res.data?.delegations?.nodes.some((i) => i?.indexer?.active === false)) {
-      notificationStore.addNotification({
-        key: NotificationKey.InactiveOperator,
-        level: 'critical',
-        message:
-          'This node operator has unregistered from SubQuery Network and you are receiving no more rewards. You should redelegate your SQT to another Node Operator to continue to receive rewards.',
-        title: 'Node Operator Inactive',
-        createdAt: Date.now(),
-        canBeDismissed: true,
-        dismissTime: 1000 * 60 * 60 * 24,
-        dismissTo: undefined,
-        type: '',
-        buttonProps: {
-          label: 'Change Delegation',
-          navigateHref: '/delegator/my-delegation',
-        },
-      });
-      notificationStore.sortNotificationList();
-    }
-  }, [account, notificationStore.notificationList]);
-
-  const makeLowControllerBalanceNotification = useCallback(async () => {
-    if (notificationStore.notificationList.find((item) => item.key === NotificationKey.LowControllerBalance)) {
-      return;
-    }
-    const res = await fetchIndexerController({
-      variables: {
-        address: account || '',
-      },
-      fetchPolicy: 'network-only',
-    });
-
-    if (res.data?.indexer?.controller) {
-      const controller = res.data.indexer.controller;
-      const controllerBalance = await provider?.getBalance(controller);
-
-      if (BigNumberJs(formatEther(controllerBalance)).lt(0.001)) {
-        notificationStore.addNotification({
-          key: NotificationKey.LowControllerBalance,
-          level: 'critical',
-          message:
-            'Your Controller Account’s balance is running now. Top up now to ensure rewards are paid out in time.',
-          title: 'Low Controller Balance',
-          createdAt: Date.now(),
-          canBeDismissed: false,
-          dismissTime: 1000 * 60 * 60 * 24,
-          dismissTo: undefined,
-          type: '',
-          buttonProps: {
-            label: 'Add Balance',
-            navigateHref: '',
-          },
-        });
-        notificationStore.sortNotificationList();
-      }
-    }
-  }, [account, notificationStore.notificationList, provider]);
-
-  const makeUnlockWithdrawalNotification = useCallback(async () => {
-    if (
-      notificationStore.notificationList.find((item) => item.key === NotificationKey.UnlockWithdrawal) ||
-      !contracts
-    ) {
-      return;
-    }
-    const lockPeriod = await limitContract(() => contracts.staking.lockPeriod(), makeCacheKey('lockPeriod'), 0);
-    const res = await fetchUnlockWithdrawal({
-      variables: {
-        address: account || '',
-        startTime: dayjs().subtract(convertBigNumberToNumber(lockPeriod), 'seconds').toISOString(),
-      },
-      fetchPolicy: 'network-only',
-    });
-
-    if (res.data?.withdrawls.totalCount) {
-      notificationStore.addNotification({
-        key: NotificationKey.UnlockWithdrawal,
-        level: 'info',
-        message: 'Your withdrawal has been unlocked.',
-        title: 'Unlock Withdrawal',
-        createdAt: Date.now(),
-        canBeDismissed: true,
-        dismissTime: 1000 * 60 * 60 * 24,
-        dismissTo: undefined,
-        type: '',
-        buttonProps: {
-          label: 'Withdraw',
-          navigateHref: '/profile/withdrawn',
-        },
-      });
-    }
-  }, [account, notificationStore.notificationList, contracts]);
-
-  const initAllNotification = useCallback(() => {
-    idleCallback(makeOverAllocateAndUnStakeAllocationNotification);
-    idleCallback(makeUnClaimedNotification);
-    idleCallback(makeLowBillingBalanceNotification);
-    idleCallback(makeInactiveOperatorNotification);
-    idleCallback(makeLowControllerBalanceNotification);
-    idleCallback(makeUnlockWithdrawalNotification);
-  }, [
-    makeOverAllocateAndUnStakeAllocationNotification,
-    makeUnClaimedNotification,
-    makeLowBillingBalanceNotification,
-    makeInactiveOperatorNotification,
-    makeLowControllerBalanceNotification,
-    makeUnlockWithdrawalNotification,
-  ]);
-
-  return {
-    makeOverAllocateNotification: () => idleCallback(makeOverAllocateAndUnStakeAllocationNotification),
-    makeUnClaimedNotification: () => idleCallback(makeUnClaimedNotification),
-    makeLowBillingBalanceNotification: () => idleCallback(makeLowBillingBalanceNotification),
-    makeInactiveOperatorNotification: () => idleCallback(makeInactiveOperatorNotification),
-    makeLowControllerBalanceNotification: () => idleCallback(makeLowControllerBalanceNotification),
-    makeUnlockWithdrawalNotification: () => idleCallback(makeUnlockWithdrawalNotification),
-    initNewNotification: () => idleCallback(initAllNotification),
-  };
 };
 
 const NotificationItem: FC<{ item: NotificationItemType }> = ({ item }) => {
@@ -575,10 +114,12 @@ const NotificationCentre: FC = () => {
     await notificationStore.initNotification(account || '');
   };
 
+  // initNewNotification must wait init finish.
+  // react's update flow need to wait until next tick or pass as parameter.
+  // I don't want to pass, just use useEffect to make sure init finish.
   useEffect(() => {
     init();
   }, [account]);
-
   useEffect(() => {
     if (notificationStore.mounted && !currentEra.loading && !currentEra.error) {
       initNewNotification();
